修改crc，添加json格式的封装发送

+++ lib/crc16.h
@@ -7,7 +7,7 @@ extern "C" {
 #endif /* __cplusplus  */
 
 unsigned short crc16_ccitt(const unsigned char *buf, int len);
-int check(int crc, const unsigned char *buf, int sz);
+int check(unsigned short crc, const unsigned char *buf, int sz);
 
 #ifdef __cplusplus
 }
